pyproject.toml
[tool.poetry]
name = "coloring-book-generator"
version = "0.1.0"
description = ""
authors = ["Felipe Pimentel <fpimentel88@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"


customtkinter = "^5.2.2"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"
transformers = "^4.45.1"
torch = "^2.4.1"
torchvision = "^0.19.1"
scikit-image = "^0.24.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "coloring_book_generator.app:main"